CLN : literalinclude configure_traits_view_group
	new file:   docs/source/traitsui_user_manual/examples/configure_traits_view_group.py
	modified:   docs/source/traitsui_user_manual/view.rst
	deleted:    examples/tutorials/doc_examples/examples/configure_traits_view_group.py
