+++ docs/source/traitsui_user_manual/view.rst
@@ -349,34 +349,8 @@ Consider the following enhancement to Example 2:
 
 .. rubric:: Example 3: Using configure_traits() with a View and a Group object
 
-::
-
-    # configure_traits_view_group.py -- Sample code to demonstrate
-    # configure_traits()
-    from traits.api import HasTraits, Int, Str
-    from traitsui.api import Group, Item, View
-    import traitsui
-
-    class SimpleEmployee(HasTraits):
-        first_name = Str()
-        last_name = Str()
-        department = Str()
-
-        employee_number = Str()
-        salary = Int()
-
-    view1 = View(
-        Group(
-            Item(name='first_name'),
-            Item(name='last_name'),
-            Item(name='department'),
-            label='Personnel profile',
-            show_border=True,
-        ),
-    )
-
-    sam = SimpleEmployee()
-    sam.configure_traits(view=view1)
+.. literalinclude:: examples/configure_traits_view_group.py
+   :start-at: configure_traits_view_group.py
 
 The resulting window shows the same widgets as before, but they are now enclosed
 in a visible border with a text label:
